requirements.txt
azure-cognitiveservices-speech==1.35.0
ibm-cloud-sdk-core==3.19.2
ibm-watson==8.0.0
python-decouple==3.8
requests==2.31.0
